+++ src/model/attribute/LayerItem.cpp
