+++ src/model/framebuffer/RGBFramebufferModel.cpp
